Ready for version v1.7.0 beta release

+++ whisperjav/config/components/asr/kotoba_faster_whisper.py
@@ -172,7 +172,7 @@ class KotobaFasterWhisperASRComponent(ASRComponent):
 
     # === Compute ===
     default_device = "cuda"
-    default_compute_type = "float32"
+    default_compute_type = "float16"
 
     # === Schema ===
     Options = KotobaFasterWhisperOptions
